Moved method to user service
