--- src/main/java/com/epam/lab/paymentsystem/utility/CurrentUser.java
@@ -1,5 +1,0 @@
-package com.epam.lab.paymentsystem.utility;
-
-public interface CurrentUser {
-  String getCurrentUserLogin();
-}
